Addressed comments on PR
