--- heatmap-service/sql/end_date.sql
@@ -1,1 +1,0 @@
-SELECT * FROM cmr WHERE dataset LIKE ($1) ORDER BY time DESC LIMIT 1;
