--- heatmap-service/sql/fetch_dataset.sql
@@ -1,2 +1,0 @@
-SELECT * FROM cmr
-WHERE dataset LIKE ($1);
